pyproject.toml
[project]
name = "foxypack-instagram-instagrapy"
version = "0.0.2"
description = "Default template for PDM package"
authors = [
    {name = "", email = "shapranov.work@gmail.com"},
]
dependencies = ["foxypack>=0.0.11", "instagrapi @ git+https://github.com/subzeroid/instagrapi.git@dependabot/pip/pydantic-2.12.5", "pandas>=2.3.3", "openpyxl>=3.1.5", "foxy-entities>=0.0.3"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "ruff>=0.14.0",
    "mypy>=1.19.0"
]
test = [
    "pytest>=9.0.2",
    "python-dotenv>=1.0.1"
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["twikitminifix.*"]
ignore_missing_imports = true